requirements.txt
robotframework
requests
robotframework-requests